[v0.5.0][eventually] refactor: make aggregate::Root easier to use
